chore: refactor layout and embed
