+++ src/app/bot-firewall/embed/page.tsx
@@ -5,5 +5,5 @@ import BotFirewall from '../BotFirewall';
 export const metadata = generateUseCaseMetadata(USE_CASES.botFirewall);
 
 export default function BotFirewallPage() {
-  return <BotFirewall />;
+  return <BotFirewall embed={true} />;
 }
