+++ src/app/layout.tsx
@@ -1,4 +1,5 @@
-import { LayoutUI } from './LayoutUI';
+import { ThirdPartyIntegrations } from '../client/thirdParty/ThirdPartyIntegrations';
+import { IS_PRODUCTION } from '../envShared';
 import Providers from './Providers';
 
 export const metadata = {
@@ -12,7 +13,8 @@ export default function RootLayout({ children }: { children: React.ReactNode }) 
     <html lang='en'>
       <body>
         <Providers>
-          <LayoutUI>{children}</LayoutUI>
+          {IS_PRODUCTION ? <ThirdPartyIntegrations /> : null}
+          {children}
         </Providers>
       </body>
     </html>
