+++ src/app/sms-pumping/embed/page.tsx
@@ -5,5 +5,5 @@ import { SmsPumpingUseCase } from '../SmsPumping';
 export const metadata = generateUseCaseMetadata(USE_CASES.smsPumping);
 
 export default function SmsPumping() {
-  return <SmsPumpingUseCase />;
+  return <SmsPumpingUseCase embed={true} />;
 }
