+++ src/client/components/MobileNavbar/MobileNavbar.tsx
@@ -12,9 +12,14 @@ import { TEST_IDS } from '../../testIDs';
 interface MobileNavbarProps {
   darkMode?: boolean;
   closeMobileMenu: () => void;
+  onReset?: () => void;
 }
-export default function MobileNavbar({ darkMode, closeMobileMenu }: MobileNavbarProps) {
-  const { mutate, isLoading: isResetLoading, shouldDisplayResetButton } = useReset({});
+export default function MobileNavbar({ darkMode, closeMobileMenu, onReset }: MobileNavbarProps) {
+  const {
+    mutate: resetScenarios,
+    isLoading: isResetLoading,
+    shouldDisplayResetButton,
+  } = useReset({ onSuccess: onReset });
 
   return (
     <nav className={classNames(styles.nav, { [styles.darkNavHeader]: darkMode })}>
@@ -23,7 +28,7 @@ export default function MobileNavbar({ darkMode, closeMobileMenu }: MobileNavbar
           <div className={classNames(styles.links, styles.top)}>
             <Button
               className={classNames(styles.resetButton, isResetLoading && styles.loading)}
-              onClick={() => mutate()}
+              onClick={() => resetScenarios()}
               variant='primary'
               outlined
               disabled={isResetLoading}
